CR-555 Contention exception no longer reaching event reciever
