+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/cloud/FirestoreDataStore.java
@@ -1,10 +1,5 @@
 package uk.gov.ons.ctp.integration.rhsvc.cloud;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.Random;
-import java.util.stream.Collectors;
-import org.springframework.stereotype.Service;
 import com.godaddy.logging.Logger;
 import com.godaddy.logging.LoggerFactory;
 import com.google.api.core.ApiFuture;
@@ -17,6 +12,10 @@ import com.google.cloud.firestore.QueryDocumentSnapshot;
 import com.google.cloud.firestore.QuerySnapshot;
 import com.google.cloud.firestore.WriteResult;
 import io.grpc.Status;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+import org.springframework.stereotype.Service;
 import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.common.error.CTPException.Fault;
 
@@ -62,7 +61,7 @@ public class FirestoreDataStore implements CloudDataStore {
       result.get();
       log.with(schema).with(key).debug("Firestore save completed");
     } catch (AbortedException e) {
-      if (e.getMessage().contains(Status.ABORTED.getCode().name()) 
+      if (e.getMessage().contains(Status.ABORTED.getCode().name())
           && e.getMessage().contains("Too much contention")) {
         // Firestore is overloaded. Use Spring exponential backoff to force a retry
         log.with("schema", schema).with("key", key).debug("Firestore contention detected", e);
