+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/event/impl/CaseEventReceiverImpl.java
@@ -11,7 +11,6 @@ import org.springframework.integration.annotation.ServiceActivator;
 import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.common.event.model.CaseEvent;
 import uk.gov.ons.ctp.common.event.model.CollectionCase;
-import uk.gov.ons.ctp.integration.rhsvc.cloud.DataStoreContentionException;
 import uk.gov.ons.ctp.integration.rhsvc.event.CaseEventReceiver;
 import uk.gov.ons.ctp.integration.rhsvc.repository.RespondentDataRepository;
 
@@ -33,12 +32,9 @@ public class CaseEventReceiverImpl implements CaseEventReceiver {
    *
    * @param caseEvent CaseEvent message from Response Management
    * @throws CTPException something went wrong
-   * @throws DataStoreContentionException if the data store is overloaded and rejected the object
-   *     until the backoff was exhausted.
    */
   @ServiceActivator(inputChannel = "acceptCaseEvent")
-  public void acceptCaseEvent(CaseEvent caseEvent)
-      throws CTPException, DataStoreContentionException {
+  public void acceptCaseEvent(CaseEvent caseEvent) throws CTPException {
 
     String caseTransactionId = caseEvent.getEvent().getTransactionId();
