+++ src/main/java/uk/gov/ons/ctp/integration/rhsvc/repository/RespondentDataRepository.java
@@ -5,15 +5,13 @@ import java.util.Optional;
 import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.common.event.model.CollectionCase;
 import uk.gov.ons.ctp.common.event.model.UAC;
-import uk.gov.ons.ctp.integration.rhsvc.cloud.DataStoreContentionException;
 
 /** Repository for Respondent Data */
 public interface RespondentDataRepository {
 
-  void writeUAC(UAC uac) throws CTPException, DataStoreContentionException;
+  void writeUAC(UAC uac) throws CTPException;
 
-  void writeCollectionCase(CollectionCase collectionCase)
-      throws CTPException, DataStoreContentionException;
+  void writeCollectionCase(CollectionCase collectionCase) throws CTPException;
 
   Optional<UAC> readUAC(String universalAccessCode) throws CTPException;
