+++ src/test/java/uk/gov/ons/ctp/integration/rhsvc/message/impl/UacEventReceiverImplUnit_Test.java
@@ -9,7 +9,6 @@ import uk.gov.ons.ctp.common.event.model.Header;
 import uk.gov.ons.ctp.common.event.model.UAC;
 import uk.gov.ons.ctp.common.event.model.UACEvent;
 import uk.gov.ons.ctp.common.event.model.UACPayload;
-import uk.gov.ons.ctp.integration.rhsvc.cloud.DataStoreContentionException;
 import uk.gov.ons.ctp.integration.rhsvc.event.impl.UACEventReceiverImpl;
 import uk.gov.ons.ctp.integration.rhsvc.repository.RespondentDataRepository;
 import uk.gov.ons.ctp.integration.rhsvc.repository.impl.RespondentDataRepositoryImpl;
@@ -29,7 +28,7 @@ public class UacEventReceiverImplUnit_Test {
   }
 
   @Test
-  public void test_acceptUACEvent_success() throws CTPException, DataStoreContentionException {
+  public void test_acceptUACEvent_success() throws CTPException {
 
     // Construct UACEvent
     UACEvent uacEventFixture = new UACEvent();
